Revert "Convert webpack.config.js into TypeScript"
