+++ package.json
@@ -18,12 +18,7 @@
   "dependencies": {
     "@fortawesome/fontawesome-free": "^5.15.4",
     "@popperjs/core": "^2.9.3",
-    "@types/compression-webpack-plugin": "^6.0.6",
     "@types/glob": "^7.1.4",
-    "@types/js-yaml": "^4.0.3",
-    "@types/mini-css-extract-plugin": "^2.2.0",
-    "@types/webpack-assets-manifest": "^5.0.2",
-    "@types/webpack-dev-server": "^4.1.0",
     "@types/zxcvbn": "^4.4.1",
     "bootstrap": "^5.1.0",
     "compression-webpack-plugin": "^8.0.1",
@@ -40,7 +35,6 @@
     "sass-loader": "^12.1.0",
     "style-loader": "^3.2.1",
     "ts-loader": "^9.2.5",
-    "ts-node": "^10.2.1",
     "typescript": "^4.3.5",
     "webpack": "^5.50.0",
     "webpack-assets-manifest": "^5.0.6",
