+++ tsconfig.json
@@ -4,12 +4,5 @@
     "target": "es2017",
     "module": "esNext",
     "esModuleInterop": true,
-    "moduleResolution": "node"
-  },
-
-  "ts-node": {
-    "compilerOptions": {
-      "module": "CommonJS"
-    }
   }
 }
